+++ examples/cpu/matmult.cpp
@@ -9,8 +9,8 @@ int main()
   cpumat<float> x(n, n);
   cpumat<float> y(n, n);
   
-  x.fill_linspace(1.f, (float) n*n);
-  y.fill_linspace((float) n*n, 1.f);
+  x.fill_linspace(1, n*n);
+  y.fill_linspace(n*n, 1);
   
   cpumat<float> z = linalg::matmult(false, false, 1.0f, x, y);
   z.print();
